requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
piexif==1.1.3
boto3==1.35.20
botocore==1.35.20
python-dotenv==1.0.1
python-multipart==0.0.9
firebase-admin==6.6.0
httpx==0.28.1
websockets==14.2
Wand==0.6.13
opencv-python-headless==4.10.0.84
numpy==1.26.4
jinja2==3.1.4
google-auth==2.34.0
mediapipe==0.10.14

# ML / CV (only those actually used)
PyJWT==2.9.0
bcrypt==4.2.0
torch==2.8.0
kornia==0.7.3

# Background removal
rembg==2.0.57
onnxruntime-gpu==1.18.1
segment-anything==1.0
timm==0.9.16

# LUT generation
PyLUT==1.4.9

# Copyright Defense Dependencies
pydantic[email]==2.8.2
cryptography==42.0.8

# Utilities used by scripts and webhooks
gdown==5.2.0
standardwebhooks==1.0.0
qrcode==7.4.2

# Rate limiting
throttled-py==2.2.3
redis==5.0.1
